Make GPT use base prompt learning model class:
